+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### [Added]
 
 * PR #580: Added features to load videos from S3.
+* PR #568: feat: added support for INSERT and DELETE operators for structured data
 
 ### [Changed]